base library

+++ Haze/Haze/Library/Header/HazeMemoryLib.h
@@ -9,5 +9,5 @@ public:
 
 	HAZE_STD_LIB_FUNCTION(void, MemoryCopy, void* dst, const void* src, int64 size);
 
-	HAZE_STD_LIB_FUNCTION(void, ObjectArrayConstructor, void* address, void* constructorAddress, int objectSize, int count);
+	HAZE_STD_LIB_FUNCTION(void, ObjectArrayConstructor, void* address, void* constructorAddress, uint64 objectSize, uint64 count);
 };
